added basic boilerplate for node
