+++ UI/public/manifest.json
