+++ package.json
@@ -1,27 +1,25 @@
 {
-  "name": "reactorcore",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "bootstrap": "^4.1.3",
-    "react": "^16.6.3",
-    "react-dom": "^16.6.3",
-    "react-scripts": "2.1.1",
-    "reactstrap": "^6.5.0"
-  },
+  "name": "reactor",
+  "version": "1.0.0",
+  "description": "Reactor: React Component Packages for Developers",
+  "main": "server.js",
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "eslintConfig": {
-    "extends": "react-app"
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/ElementalXriath/Reactor.git"
   },
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 11",
-    "not op_mini all"
-  ]
+  "author": "Johnny Moore & Joseph Thompson",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/ElementalXriath/Reactor/issues"
+  },
+  "homepage": "https://github.com/ElementalXriath/Reactor#readme",
+  "dependencies": {
+    "body-parser": "^1.18.3",
+    "cors": "^2.8.5",
+    "express": "^4.16.4",
+    "passport": "^0.4.0"
+  }
 }
