+++ server.js
@@ -1,0 +1,7 @@
+const express = require('express');
+
+const server = express();
+
+server.listen('3000', () => {
+  console.log('Running on PORT:3000');
+})
